% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listcomp.R
\name{lc}
\alias{lc}
\title{Basic list comprehensions in R.}
\usage{
lc(string)
}
\arguments{
\item{string}{text denoting the list comprehension to evaluate}
}
\value{
evaluated list comprehension
}
\description{
Pass a Python-style list comprehension as a string and have it evaluated in
the calling environment.
Supports hopefully anything with complexity less than or equal to: \cr
 `fn_a(item) for item in my_list if fn_b(item) == TRUE` \cr
and greater than or equal to: \cr
 `item for item in list` \cr
Supports nesting by wrapping inner comprehensions with `[]`.
}
\examples{
my_sequence <- seq(10)
nested_list <- list(seq(5), seq(10))
lc('item ^ 2 for item in my_sequence if item \%\% 2 == 0')
lc('x ^ 2 for x in [max(y) for y in nested_list]')

library("ggplot2")
data(diamonds)
lc('mean(x) for x in diamonds if is.numeric(x)')
}
\author{
Dex Groves
}

