% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listcomp.R
\name{listcomp}
\alias{listcomp}
\title{Basic list comprehensions in R.}
\usage{
listcomp(string)
}
\arguments{
\item{string}{text denoting the list comprehension to evaluate}
}
\value{
evaluated list comprehension
}
\description{
Pass a Python-style list comprehension as a string and have it evaluated in
the calling environment.
Supports hopefully anything as long as its complexity is less than or equal
to: \cr
 `fn_a(item) for item in my_list if fn_b(item) == TRUE` \cr
and greater than or equal to: \cr
 `item for item in list` \cr
Does not support nesting.
}
\author{
Dex Groves
}

