% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst.R
\name{lst}
\alias{lst}
\title{List comprehensions}
\usage{
lst(element_expr, ..., .compile = TRUE)
}
\arguments{
\item{element_expr}{an expression that will be collected}

\item{...}{either logical expression or a named parameter with an iterable
sequence.}

\item{.compile}{compile the resulting for loop to bytecode befor eval}
}
\description{
Syntactic sugar for for loops
}
\examples{
lst(c(x, y), x = 1:10, y = 1:10, x + y == 10, x < y)
z <- 10
lst(c(x, y), x = 1:10, y = 1:10, x + y == !!z, x < y)

}
